+++ js/interfaces/formData.ts
@@ -1,0 +1,3 @@
+export interface FormSerialisable<T> {
+    deserialise(formData: any): T;
+}
