+++ ui-debug.log
@@ -1,1 +1,1 @@
-Web / API server started at http://localhost:4004
+Web / API server started at http://localhost:4000
